Avoid reading entire file for tag identification
ID3v1 (will be added in a future pr) tag identifier is located at the end of the file.
The current strategy of identifying the tag type is to merge all range identifiers and make a single request.
With ID3v1 this would lead to reading the entire file. Avoid it by making two requests, one for start of file tag identifiers and another for end of file.