Fix bug in ChunkedFileData with overlapping chunk
When appending data to the overlapping chunk at the tail the wrong data array was being used.
This was only an issue when there was also overlapping data at the head (otherwise both data arrays would be the same).

Also fixed the way we check if the last chunk was needed when appending data. Not an issue but useless operations were being made.