Make XHR requests range in multiples of 1K
The bottleneck of small range requests is establishing the network connection. For this reason we might as well just round the range to 1KB multiples and get more data while we're at it.

Tags are usually bound together so it's a waste of network requests if we only get the specified data.

On an example MP4 file I was able to reduce the number of network requests from 8 to 2.