Clean up variable naming and add more comments
Some of the variables were named in ways that were difficult to
understand, so a couple have been renamed to make their purpose a little
bit more clear.

In addition, a lot of new comment blocks have been added, which should
hopefully explain more of the FLAC specification and the purposes of all
the different numbers.